Split storage module in a separate load and save modules
