+++ gaphor/extensions/sphinx.py
@@ -12,11 +12,11 @@ from docutils.parsers.rst.directives import images
 from sphinx.util import logging
 from sphinx.util.typing import OptionSpec
 
+import gaphor.storage as storage
 from gaphor.core.modeling import Diagram, ElementFactory
 from gaphor.diagram.export import save_pdf, save_svg
 from gaphor.i18n import gettext
 from gaphor.services.modelinglanguage import ModelingLanguageService
-from gaphor.storage import storage
 
 log = logging.getLogger(__name__)
